requirements.txt
git+https://github.com/wbaek/theconf
git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git
git+https://github.com/ildoonet/pystopwatch2.git

keras==2.4.0
tensorflow-datasets==4.3.0
tensorflow-probability==0.13.0
matplotlib
seaborn
pandas
packaging

colored
pretrainedmodels
tqdm
tensorboardx
sklearn
matplotlib
psutil
requests
Pillow